DMP-2623 renamed 'template' package to 'model'
