+++ src/main/java/uk/gov/hmcts/darts/casedocument/mapper/CourtCaseDocumentMapper.java
@@ -6,9 +6,9 @@ import org.mapstruct.Mappings;
 import org.mapstruct.Named;
 import org.springframework.beans.factory.annotation.Autowired;
 import uk.gov.hmcts.darts.authorisation.component.UserIdentity;
-import uk.gov.hmcts.darts.casedocument.template.CaseRetentionCaseDocument;
-import uk.gov.hmcts.darts.casedocument.template.CourtCaseDocument;
-import uk.gov.hmcts.darts.casedocument.template.HearingCaseDocument;
+import uk.gov.hmcts.darts.casedocument.model.CaseRetentionCaseDocument;
+import uk.gov.hmcts.darts.casedocument.model.CourtCaseDocument;
+import uk.gov.hmcts.darts.casedocument.model.HearingCaseDocument;
 import uk.gov.hmcts.darts.common.entity.CaseManagementRetentionEntity;
 import uk.gov.hmcts.darts.common.entity.CourtCaseEntity;
 import uk.gov.hmcts.darts.common.entity.HearingEntity;
