+++ src/main/java/uk/gov/hmcts/darts/casedocument/model/DefendantCaseDocument.java
@@ -1,0 +1,12 @@
+package uk.gov.hmcts.darts.casedocument.model;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class DefendantCaseDocument extends CreatedModifiedCaseDocument {
+
+    private final Integer id;
+    private final String name;
+}
