+++ src/main/java/uk/gov/hmcts/darts/casedocument/model/ProsecutorCaseDocument.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.darts.casedocument.template;
+package uk.gov.hmcts.darts.casedocument.model;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class DefenceCaseDocument extends CreatedModifiedCaseDocument {
+public class ProsecutorCaseDocument extends CreatedModifiedCaseDocument {
 
     private final Integer id;
     private final String name;
