+++ src/main/java/uk/gov/hmcts/darts/casedocument/service/CaseDocumentService.java
@@ -1,6 +1,6 @@
 package uk.gov.hmcts.darts.casedocument.service;
 
-import uk.gov.hmcts.darts.casedocument.template.CourtCaseDocument;
+import uk.gov.hmcts.darts.casedocument.model.CourtCaseDocument;
 
 public interface CaseDocumentService {
